requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
requests==2.27.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
